Added configurable margins.

+++ Editor/src/main/java/se/natusoft/doc/markdowndoc/editor/config/IntegerConfigEntry.java
@@ -99,14 +99,14 @@ public class IntegerConfigEntry extends ConfigEntry {
     /**
      * Returns the minimum value.
      */
-    public double getMinValue() {
+    public int getMinValue() {
         return this.min;
     }
 
     /**
      * Returns the maximum value.
      */
-    public double getMaxValue() {
+    public int getMaxValue() {
         return this.max;
     }
 }
